Introduce Try type to preserve failure cause
Prior to this commit, the cause for a failure to load a class was not
available anymore when throwing an exception at a later point in time
because `ReflectionUtils` returned an empty `Optional<Class<?>>`. The
new `Try` type encapsulates either a successful invocation with a
(nullable) resulting value or a failure with an exception as its cause.